Rebase media-libs/mesa-9999 on x11 overlay's
FireBurn is dead, this (if it continues to exist) needs to be migrated
to robinkverlay.  iXit's repo is sort of quiet right now, I don't need a
spiffy D3D9 state tracker at the moment, and all the exiciting Radeon
VCE and OpenMAX work is happening in fd.o Mesa master.  Changed the name
to -r1, too, since =media-libsa/mesa-9999 has GX86 multilib support.
So, yeah...